Revert compatibility changes and fix small issues
